requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
emoji==2.14.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
omegaconf==2.3.0
packaging==25.0
portalocker==3.1.1
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyparsing==3.2.3
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
stanza==1.10.1
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
torch-geometric==2.6.1
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
